remove isFunctionHigherOrder from common utils
